crud product

+++ backend/controllers/ProductController.mjs
@@ -1,5 +1,7 @@
+import { Sequelize } from "sequelize";
+
 // import models
-import ProductModel from "../models/Product.mjs";
+import ProductModel from "../models/product.js";
 
 // import helpers
 import FormatResponse from "../helpers/FormatResponse.mjs";
@@ -7,43 +9,34 @@ import FormatResponse from "../helpers/FormatResponse.mjs";
 // import validators
 import {ValidateProduct} from "../validators/ProductValidator.mjs";
 
-import { client } from "../elastic/connection.mjs";
-
 import fs from "fs";
 
 // function get All Products
 export const ProductIndex = async (req, res) => {
     try {
-       let filter = {};
+        let filter = {
+            order: [
+                [req.query.order || 'id', (req.query.ordering === 'asc' ? 'asc' : 'desc')],
+            ],
+            attributes : ['id','title','stock','price'],
+            limit : req.query.per_page || 10
+       };
 
        if(req.query.search){
-            filter = {
-                $and : [{
-                    $or : [
-                        {
-                            title : {
-                                $regex : ".*" + req.query.search + ".*",
-                            },                                                        
-                        }
-                    ]
-                }]           
+            filter.where = {            
+                    [Sequelize.Op.and] : [{
+                        [Sequelize.Op.or] : [
+                            {
+                                title : {
+                                    [Sequelize.Op.regexp] : ".*" + req.query.search + ".*",
+                                },                                          
+                            },                           
+                        ]
+                    }]                           
             }
-       }
+       }    
 
-       const products = await ProductModel
-        .find(filter)
-        .limit(req.query.per_page || 5)
-        .sort({
-            [req.query.order || 'createdAt'] : (req.query.ordering === 'asc' ? 1 : -1)
-        })
-        .select({
-            _id : 1,
-            title : 1,
-            price : 1,
-            photo : 1.,
-            description : 1,
-            stock : 1
-        });
+       const products = await ProductModel.findAll(filter);
 
        return res.json(products);
     } catch (error) {       
@@ -56,13 +49,12 @@ export const ProductShow = async (req, res) => {
     try {
 
         const product = await ProductModel
-            .findById(req.params.id)
-            .select({
-                _id : 1,
-                title : 1,
-                price : 1,
-                photo : 1
-            });
+            .findOne({
+                where : {
+                    id: req.params.id
+                },
+                attributes : ['id','title','stock','price'],
+            })    
 
         if(!product){
             return res.status(401).json({
@@ -73,8 +65,7 @@ export const ProductShow = async (req, res) => {
         return res.json(product);
     } catch (error) {
         return FormatResponse.Failed(error,res);
-    }
-     
+    }    
 }
  
 export const ProductCreate = async (req, res) => {
@@ -91,30 +82,7 @@ export const ProductCreate = async (req, res) => {
             req.body.photo = req.file.filename; 
         }
     
-        let resMongodb = await new ProductModel(req.body).save()    
-
-        console.log(resMongodb);
-
-        console.log(resMongodb.id);
-        console.log(resMongodb.title);
-        console.log(resMongodb.price);
-        console.log(resMongodb.stock);
-        console.log(resMongodb.description);
-               
-        let resElastic = await client.create({
-          index: 'products',
-          type: 'product',
-          id : resMongodb.id,
-          body: {
-                title : resMongodb.title,
-                price : resMongodb.price,
-                stock : resMongodb.stock,
-                description : resMongodb.description,
-                createdAt : resMongodb.createdAt
-            }          
-        })
-
-        console.log(resElastic);
+        await ProductModel.create(req.body)
 
         return res.json({
             "message" : true
@@ -136,11 +104,12 @@ export const ProductUpdate = async (req, res) => {
         }    
 
         const product = await ProductModel          
-            .findById(req.params.id)
-            .select({
-                _id : 1,
-                photo : 1
-            });
+            .findOne({
+                where : {
+                    id : req.params.id
+                },  
+                attributes : ['id','photo']
+            });  
 
         if(!product) {            
             return res.status(404).json({
@@ -151,34 +120,16 @@ export const ProductUpdate = async (req, res) => {
         if(req.file){                        
             req.body.photo = req.file.filename;             
 
-            if(product.photo && fs.existsSync("./assets/products/"+product.photo)){
+            if(product.photo && product.photo != "default.png" && fs.existsSync("./assets/products/"+product.photo)){
                 fs.unlinkSync("./assets/products/"+product.photo);
             }
         }
 
-        let resMongodb = await ProductModel.updateOne({
-            _id: req.params.id
-        }, {
-            $set: req.body
-        });
-
-        console.log(resMongodb);
-
-        let resElastic = await client.update({
-            index: "products",
-            type: "product",
-            id: req.params.id,
-            body: {            
-                doc: {
-                    price: req.body.price,
-                    stock: req.body.stock,
-                    title: req.body.title,
-                    description : req.body.description
-                }
+        await ProductModel.update(req.body, {
+            where: {
+              id: req.params.id
             }
-        })
-
-        console.log(resElastic);
+        });
 
         return res.json({
             message : true
@@ -192,11 +143,12 @@ export const ProductUpdate = async (req, res) => {
 export const ProductDestroy = async (req, res) => {
     try {        
         const product = await ProductModel           
-            .findById(req.params.id)
-            .select({
-                _id : 1,
-                photo : 1
-            });
+            .findOne({
+                where : {
+                    id : req.params.id
+                },  
+                attributes : ['id','photo']
+            });    
 
         if(!product) {            
             return res.status(404).json({
@@ -204,23 +156,15 @@ export const ProductDestroy = async (req, res) => {
             });
         }
 
-        if(product.photo && fs.existsSync("./assets/products/"+product.photo)){
+        if(product.photo && product.photo != "default.png" && fs.existsSync("./assets/products/"+product.photo)){
             fs.unlinkSync("./assets/products/"+product.photo);
         }
 
-        let resMongodb = await ProductModel.deleteOne({
-            _id: req.params.id
+        await ProductModel.destroy({
+            where : {
+                id: req.params.id
+            }
         });
-        
-        console.log(resMongodb);
-
-        let resElastic = await client.delete({
-            index: "products",
-            type: "product",
-            id: req.params.id
-        })
-
-        console.log(resElastic);
 
         return res.json({
             message : true
@@ -228,48 +172,4 @@ export const ProductDestroy = async (req, res) => {
     } catch (error) {    
         return FormatResponse.Failed(error,res);
     }
-}
-
-// searching -> done
-// limit -> done
-// orderBy -> done
-// select -> done
-
-// function Search Product With ElasticSearch
-export const ProductElastic =  async (req,res) => {
-   try {
-    let initialSearch = {
-        index: 'products',
-        type: 'product',
-        body: {         
-            // _source : ["title"], -> only select
-            // from : 1,
-            size : req.query.per_page || 5,        
-            sort : [
-                {"createdAt" : "desc"}                            
-            ]
-        }        
-    };
-
-    if(req.query.search){        
-        initialSearch.body.query = {    
-            multi_match : {
-                "query" : req.query.search,
-                "type" : "phrase",
-                "fields" : ["description","title","price","stock"]
-            }
-        }
-    }
-
-   
-    let response = await client.search(initialSearch);
-
-    console.log(response.body.hits.hits);
-
-    return res.json({
-        "message" : true
-    })
-   }catch (error) {        
-    return FormatResponse.Failed(error,res);
-   }
 }
